+++ frontend/lib/models/auth_user.dart
@@ -5,8 +5,8 @@ part 'auth_user.g.dart';
 @JsonSerializable()
 class AuthUser {
   const AuthUser({
-    required this.userId,
-    required this.account,
+    this.publicId,
+    this.userId,
     required this.email,
     required this.userName,
     required this.avatarUrl,
@@ -15,8 +15,8 @@ class AuthUser {
   factory AuthUser.fromJson(Map<String, dynamic> json) =>
       _$AuthUserFromJson(json);
 
-  final int userId;
-  final String account;
+  final String? publicId;
+  final int? userId;
   final String email;
   final String userName;
   final String? avatarUrl;